New Prefetch Controller
