+++ hw/chisel/src/main/scala/snax/readerWriter/AddressGenUnit.scala
@@ -1,92 +1,8 @@
 package snax.readerWriter
 
 import snax.utils._
-
 import chisel3._
 import chisel3.util._
-
-/** This class represents a basic counter module in Chisel.
-  *
-  * @param width
-  *   The width of the counter.
-  * @param hasCeil
-  *   Indicates whether the counter has a ceiling value.
-  */
-
-class BasicCounter(width: Int, hasCeil: Boolean = true)
-    extends Module
-    with RequireAsyncReset {
-  val io = IO(new Bundle {
-    val tick = Input(Bool())
-    val reset = Input(Bool())
-    val ceil = Input(UInt(width.W))
-
-    val value = Output(UInt(width.W))
-    val lastVal = Output(Bool())
-  })
-  // 32.W should be enough to count any loops
-  val nextValue = Wire(UInt(width.W))
-  val value = RegNext(nextValue, 0.U)
-  nextValue := {
-    if (hasCeil) {
-      Mux(
-        io.reset,
-        0.U,
-        Mux(io.tick, Mux(value < io.ceil - 1.U, value + 1.U, 0.U), value)
-      )
-    } else {
-      Mux(io.reset, 0.U, Mux(io.tick, value + 1.U, value))
-    }
-  }
-
-  io.value := value
-  io.lastVal := {
-    (if (hasCeil) (value === io.ceil - 1.U) else (value.andR)) && io.tick
-  }
-}
-
-class ProgrammableCounter(width: Int, hasCeil: Boolean = true)
-    extends Module
-    with RequireAsyncReset {
-  val io = IO(new Bundle {
-    val tick = Input(Bool())
-    val reset = Input(Bool())
-    val ceil = Input(UInt(width.W))
-    val step = Input(UInt((width - 1).W))
-
-    val value = Output(UInt(width.W))
-    val lastVal = Output(Bool())
-  })
-  val nextValue = Wire(UInt(width.W))
-  val value = RegNext(nextValue, 0.U)
-
-  // If has ceil, a small counter is used to count the step
-  // The small counter's function is to determine whether the ceil is reached, and a reset is needed.
-  if (hasCeil) {
-    val smallCounter = Module(new BasicCounter(width, hasCeil) {
-      override val desiredName = "ProgrammableCounter_SmallCounter"
-    })
-
-    smallCounter.io.tick := io.tick
-    smallCounter.io.reset := io.reset
-    smallCounter.io.ceil := io.ceil
-    io.lastVal := smallCounter.io.lastVal
-  }
-
-  nextValue := {
-    if (hasCeil) {
-      Mux(
-        io.reset || (io.lastVal && io.tick),
-        0.U,
-        Mux(io.tick, value + io.step, value)
-      )
-    } else {
-      Mux(io.reset, 0.U, Mux(io.tick, value + io.step, value))
-    }
-  }
-
-  io.value := value
-}
 
 /** AGU is the module to automatically generate the address for all ports.
   * @input
