+++ hw/chisel/src/main/scala/snax/readerWriter/Reader.scala
@@ -115,17 +115,13 @@ class Reader(
       responser.in.tcdmRsp <> tcdmRsp
     }
   }
-  // Responser <> DataBuffer, Data Link + NearlyFull
+  // Responser <> DataBuffer, Data Link + dataFifoPopped
   dataBuffer.io.in.zip(responsers.io).foreach {
     case (buffer, responser) => {
       buffer <> responser.out.data
     }
   }
-  responsers.io.zip((dataBuffer.io.nearlyFull)).foreach {
-    case (responser, nearlyFull) => {
-      responser.out.dataFifoNearlyFull := nearlyFull
-    }
-  }
+  responsers.io.foreach(_.out.dataFifoPopped := dataBuffer.io.out.head.fire)
 
   // DataBuffer <> Output
   dataBuffer.io.out.head <> io.data
