+++ hw/chisel/src/main/scala/snax/utils/ComplexQueue.scala
@@ -47,20 +47,18 @@ class ComplexQueueConcat(
       },
       Decoupled(UInt(outputWidth.W))
     )
-    val nearlyEmpty = Output(Vec(numChannel, Bool()))
-    val nearlyFull = Output(Vec(numChannel, Bool()))
     val allEmpty = Output(Bool())
     val anyFull = Output(Bool())
   })
 
   val queues = for (i <- 0 until numChannel) yield {
     val queue = Module(new Queue(UInt(smallWidth.W), depth, pipe))
-    io.nearlyEmpty(
-      i
-    ) := queue.io.count === 0.U || (queue.io.count === 1.U && ~queue.io.enq.fire)
-    io.nearlyFull(
-      i
-    ) := queue.io.count === depth.U || (queue.io.count === (depth - 1).U && ~queue.io.deq.fire)
+    // io.nearlyEmpty(
+    //   i
+    // ) := queue.io.count === 0.U || (queue.io.count === 1.U && ~queue.io.enq.fire)
+    // io.nearlyFull(
+    //   i
+    // ) := queue.io.count === depth.U || (queue.io.count === (depth - 1).U && ~queue.io.deq.fire)
     queue
   }
